database_sharing_service/requirements.txt
passlib[bcrypt]==1.7.4    # Password hashing library with bcrypt support
pydantic==2.8.2           # Data validation and settings management
pydantic[email]           # Email validation plugin for Pydantic
sqlalchemy==2.0.0         # ORM for database interaction
alembic==1.11.1           # Database migrations tool compatible with SQLAlchemy
psycopg2==2.9.9           # PostgreSQL database adapter
python-dotenv==1.0.0      # Loading environment variables from a shared .env file
#psycopg2-binary==2.9.1    # PostgreSQL database adapter